rename starfish_py folder
