+++ docs/source/classes/asset_base.rst
@@ -1,7 +1,7 @@
 AassetBase class
 ================
 
-.. automodule:: starfish_py.asset.asset_base
+.. automodule:: starfish.asset.asset_base
     :members:
     :undoc-members:
     :show-inheritance:
