+++ ocean
@@ -7,10 +7,10 @@ import sys
 import logging
 
 
-from starfish_py.command_line import CommandLine
-from starfish_py.exceptions import OceanCommandLineError
+from starfish.command_line import CommandLine
+from starfish.exceptions import OceanCommandLineError
 
-from starfish_py import Config
+from starfish import Config
 
 DEFAULT_CONFIG_FILE = 'config_local.ini'
