+++ starfish/asset/asset.py
@@ -9,14 +9,14 @@ from eth_utils import remove_0x_prefix
 
 from squid_py.did import did_to_id
 
-from starfish_py import Account
-from starfish_py.asset.asset_base import AssetBase
-from starfish_py.models.squid_model import SquidModel
+from starfish import Account
+from starfish.asset.asset_base import AssetBase
+from starfish.models.squid_model import SquidModel
 
-from starfish_py.utils.did import did_parse
+from starfish.utils.did import did_parse
 
 
-# from starfish_py import logger
+# from starfish import logger
 
 
 class Asset(AssetBase):
