+++ starfish/asset/asset_light.py
@@ -9,11 +9,11 @@ from web3 import Web3
 from squid_py.did import id_to_did
 
 from eth_utils import remove_0x_prefix
-from starfish_py.asset.asset_base import AssetBase
-from starfish_py.models.metadata_agent_model import MetadataAgentModel
-from starfish_py.utils.did import did_parse
+from starfish.asset.asset_base import AssetBase
+from starfish.models.metadata_agent_model import MetadataAgentModel
+from starfish.utils.did import did_parse
 
-# from starfish_py import logger
+# from starfish import logger
 
 class AssetLight(AssetBase):
     """
