+++ starfish/command_line/base_output.py
