+++ starfish/utils/did.py
