+++ tests/test_config.py
@@ -8,11 +8,12 @@ import unittest
 import os
 import logging
 
-from starfish_py.config import Config as OceanConfig
-from starfish_py.logging import setup_logging
-from starfish_py import logger
 from squid_py import Ocean as SquidOcean
 from squid_py import Config as SquidConfig
+
+from starfish.config import Config
+from starfish.logging import setup_logging
+from starfish import logger
 
 setup_logging(level=logging.DEBUG)
 logging.getLogger("urllib3").setLevel(logging.WARNING)
@@ -34,18 +35,18 @@ TEST_VALUES = {
 def test_config_load():
 
     test_contract_path = 'test_contract_path'
-    config = OceanConfig(contract_path = test_contract_path)
+    config = Config(contract_path = test_contract_path)
     assert(config)
     assert(config.contract_path == test_contract_path)
 
     test_contract_path_env = 'test_contract_path_env'
 
     os.environ['CONTRACT_PATH'] = test_contract_path_env
-    config = OceanConfig()
+    config = Config()
     assert(config)
     assert(config.contract_path == test_contract_path_env)
 
-    config = OceanConfig(TEST_VALUES)
+    config = Config(TEST_VALUES)
     assert(config)
     # remove environ setting
     del os.environ['CONTRACT_PATH']
@@ -53,7 +54,7 @@ def test_config_load():
 def test_config_generation_for_squid():
     # start off with the most basic info to connect to ocean using squid-py library
 
-    config = OceanConfig(
+    config = Config(
         {
             'ocean_url': 'http://localhost:8545',
             'contract_path': 'artifacts'
